Refactor the serial communication code.
